Clean-up in TLSRecvLine().
Unable to track down an uninintialized memory read in its helper
function, I simplified the code.  The old code was also rescanning all
old input, each time it appended new, looking for a '\n' that could
only be in the new; let's not waste time doing that !  If the UMR is
still there, it should be easier to find now ...

Also changed its return type to ssize_t (if there's any sense in the
buffer size being size_t, the return could overflow int) and fix its
garbled documentation.  (The @return refers to a parameter #line that
didn't exist, while talking about things other than the return value
and being wrong about whether the return count included terminating
'\0'; and the actual parameters weren't documented.)

+++ tests/unit/tls_generic_test.c
@@ -1382,11 +1382,11 @@ static void test_TLSBasicIO(void)
 
     USE_ORIGINAL(SSL_write);
     SSL_READ_RETURN(0);
-    assert_int_equal(-1, TLSRecvLine(ssl, input_buffer, output_buffer_length));
+    assert_int_equal((ssize_t)-1, TLSRecvLine(ssl, input_buffer, output_buffer_length));
     SSL_READ_RETURN(-1);
-    assert_int_equal(-1, TLSRecvLine(ssl, input_buffer, output_buffer_length));
+    assert_int_equal((ssize_t)-1, TLSRecvLine(ssl, input_buffer, output_buffer_length));
     SSL_READ_RETURN(5);
-    assert_int_equal(-1, TLSRecvLine(ssl, input_buffer, 10));
+    assert_int_equal((ssize_t)-1, TLSRecvLine(ssl, input_buffer, 10));
     SSL_READ_USE_BUFFER(output_line_buffer);
     assert_int_equal(5, TLSRecvLine(ssl, input_buffer, output_line_buffer_length));
     assert_string_equal(output_just_hello, input_buffer);
